requirements.txt
cachetools==3.0.0
certifi==2018.11.29
chardet==3.0.4
cssselect==1.0.3
dateutils==0.6.6
future==0.17.1
google-api-core==1.7.0
google-auth==1.6.2
google-cloud-core==0.29.1
google-cloud-storage==1.13.2
google-resumable-media==0.3.2
googleapis-common-protos==1.5.5
idna==2.8
lxml==4.6.5
podgen==1.0.0
protobuf==3.6.1
pyasn1==0.4.5
pyasn1-modules==0.2.3
pydub==0.23.0
pyquery==1.4.0
python-dateutil==2.7.5
pytz==2018.7
requests==2.22.0
rsa==4.7
six==1.12.0
tinytag==1.0.0
newspaper3k==0.2.8
